Create suppress output as helper
