+++ tgm/data.py
@@ -4,7 +4,7 @@ import copy
 import csv
 import pathlib
 from dataclasses import dataclass, fields, replace
-from typing import Any, Callable, List, Tuple
+from typing import Any, List, Tuple
 
 import numpy as np
 import torch
@@ -19,6 +19,7 @@ from tgm.exceptions import (
 )
 from tgm.split import SplitStrategy, TemporalRatioSplit, TGBSplit
 from tgm.timedelta import TGB_TIME_DELTAS, TimeDeltaDG
+from tgm.util._tgb import suppress_output
 
 
 @dataclass
@@ -658,34 +659,6 @@ class DGData:
             from tgb.nodeproppred.dataset import NodePropPredDataset
         except ImportError:
             raise ImportError('TGB required to load TGB data, try `pip install py-tgb`')
-
-        def suppress_output(func: Callable, *args: Any, **kwargs: Any) -> Any:
-            # This is a hacky workaround that tries to lower the verbosity on TGB
-            # logs which are currently directed to stdout. This should be removed
-            # once https://github.com/shenyangHuang/TGB/issues/117 is addressed.
-            import builtins
-
-            SILENCE_PREFIXES = [
-                'raw file found',
-                'Dataset directory is',
-                'loading processed file',
-            ]
-
-            original_print = builtins.print
-
-            def filtered_print(*p_args: Any, **p_kwargs: Any) -> None:
-                if not p_args:
-                    return
-                msg = str(p_args[0])
-                if any(msg.startswith(prefix) for prefix in SILENCE_PREFIXES):
-                    return
-                original_print(*p_args, **p_kwargs)
-
-            try:
-                builtins.print = filtered_print
-                return func(*args, **kwargs)
-            finally:
-                builtins.print = original_print
 
         if name.startswith('tgbl-'):
             dataset = suppress_output(LinkPropPredDataset, name=name, **kwargs)
